Improve handling of Back buttons, move .role calls to regular function calls, not bound ones.
Previously, when you click on a Back button, the browser tried to put the scroll to the position of the element even if it is invisible.
We incercept the click and store the position before the browser has had a chance to scroll the window.

The .role(..) function was updated to take an extra first argument which is the control being turned into a widget.
This was previously invoked as .role.call(control, role)